refactor(trading): reorganise trading detail to the correct folders
