+++ packages/suite-desktop-ui/src/support/desktopComponents.ts
@@ -32,15 +32,15 @@ import WalletSignVerify from 'src/views/wallet/sign-verify';
 import { WalletStaking } from 'src/views/wallet/staking/WalletStaking';
 import { Tokens } from 'src/views/wallet/tokens';
 import { TradingBuyConfirm } from 'src/views/wallet/trading/buy/TradingBuyConfirm';
-import { TradingBuyDetail } from 'src/views/wallet/trading/buy/TradingBuyDetail';
+import { TradingBuyDetail } from 'src/views/wallet/trading/buy/TradingBuyDetail/TradingBuyDetail';
 import { TradingBuyForm } from 'src/views/wallet/trading/buy/TradingBuyForm';
 import { TradingConciergeDetail } from 'src/views/wallet/trading/concierge/TradingConciergeDetail';
 import { TradingExchangeConfirm } from 'src/views/wallet/trading/exchange/TradingExchangeConfirm';
-import { TradingExchangeDetail } from 'src/views/wallet/trading/exchange/TradingExchangeDetail';
+import { TradingExchangeDetail } from 'src/views/wallet/trading/exchange/TradingExchangeDetail/TradingExchangeDetail';
 import { TradingExchangeForm } from 'src/views/wallet/trading/exchange/TradingExchangeForm';
 import { TradingRedirect } from 'src/views/wallet/trading/redirect/TradingRedirect';
 import { TradingSellConfirm } from 'src/views/wallet/trading/sell/TradingSellConfirm';
-import { TradingSellDetail } from 'src/views/wallet/trading/sell/TradingSellDetail';
+import { TradingSellDetail } from 'src/views/wallet/trading/sell/TradingSellDetail/TradingSellDetail';
 import { TradingSellForm } from 'src/views/wallet/trading/sell/TradingSellForm';
 import { TradingTransactions } from 'src/views/wallet/trading/transactions/TradingTransactions';
 import { Transactions } from 'src/views/wallet/transactions/Transactions';
