+++ packages/suite/src/views/wallet/trading/buy/TradingBuyDetail/TradingBuyDetailPaymentFailed.tsx
@@ -1,4 +1,4 @@
-import { type SellFiatTrade, type SellProviderInfo } from 'invity-api';
+import { type BuyProviderInfo, type BuyTrade } from 'invity-api';
 
 import { Translation } from '@suite/intl';
 import { goto } from '@suite/router';
@@ -6,49 +6,43 @@ import { Button, Card, Column, H3, IconCircle, Paragraph } from '@trezor/compone
 import { XIcon } from '@trezor/icons';
 
 import { useDispatch } from 'src/hooks/suite';
-import { type Account } from 'src/types/wallet';
+import { TradingDetailProviderInfo } from 'src/views/wallet/trading/common/TradingDetail/TradingDetailProviderInfo';
+import { TradingDetailSupportBanner } from 'src/views/wallet/trading/common/TradingDetail/TradingDetailSupportBanner';
 
-import { TradingDetailProviderInfo } from '../TradingDetailProviderInfo';
-import { TradingDetailSupportBanner } from '../TradingDetailSupportBanner';
-
-type TradingDetailSellPaymentFailedProps = {
-    account: Account;
-    trade: SellFiatTrade;
-    provider?: SellProviderInfo;
+type TradingBuyDetailPaymentFailedProps = {
+    trade: BuyTrade;
+    provider?: BuyProviderInfo;
 };
 
-export const TradingDetailSellPaymentFailed = ({
-    account,
+export const TradingBuyDetailPaymentFailed = ({
     trade,
     provider,
-}: TradingDetailSellPaymentFailedProps) => {
+}: TradingBuyDetailPaymentFailedProps) => {
     const dispatch = useDispatch();
 
-    const handleClick = () => dispatch(goto({ routeName: 'wallet-trading-sell' }));
+    const handleClick = () => dispatch(goto({ routeName: 'wallet-trading-buy' }));
 
     return (
         <Column gap={24} padding={{ top: 12, bottom: 4 }}>
             <IconCircle icon={XIcon} intent="critical" size={96} />
             <Column>
                 <H3 data-testid="@trading/transaction/detail/status">
-                    <Translation id="TR_SELL_DETAIL_ERROR_TITLE" />
+                    <Translation id="TR_BUY_DETAIL_ERROR_TITLE" />
                 </H3>
                 <Paragraph typographyStyle="body-sm" intent="neutral" priority="secondary">
-                    <Translation id="TR_SELL_DETAIL_ERROR_TEXT" />
+                    <Translation id="TR_BUY_DETAIL_ERROR_TEXT" />
                 </Paragraph>
             </Column>
             <Button onClick={handleClick} intent="neutral" priority="secondary">
-                <Translation id="TR_SELL_DETAIL_ERROR_BUTTON" />
+                <Translation id="TR_BUY_DETAIL_ERROR_BUTTON" />
             </Button>
             <Card>
                 <Column gap={24}>
                     {provider && (
                         <TradingDetailProviderInfo
-                            account={account}
-                            orderId={trade.orderId}
+                            orderId={trade.paymentId}
                             provider={provider}
                             trade={trade}
-                            txId={trade.txid}
                         />
                     )}
                     <TradingDetailSupportBanner provider={provider} trade={trade} />
