+++ packages/suite/src/views/wallet/trading/exchange/TradingExchangeDetail/TradingExchangeDetail.tsx
@@ -3,25 +3,26 @@ import { selectTradingProviderByNameAndTradeType } from '@suite-common/trading';
 import { useSelector } from 'src/hooks/suite';
 import { TradingDetailContext, useTradingDetail } from 'src/hooks/wallet/trading/useTradingDetail';
 import { TradingContainer } from 'src/views/wallet/trading/common/TradingContainer';
-import { TradingDetailSell } from 'src/views/wallet/trading/common/TradingDetail/TradingDetailSell/TradingDetailSell';
 
-export const TradingSellDetail = () => {
+import { TradingExchangeDetailContent } from './TradingExchangeDetailContent';
+
+export const TradingExchangeDetail = () => {
     const tradingDetailContext = useTradingDetail({
-        tradeType: 'sell',
+        tradeType: 'exchange',
     });
 
     const provider = useSelector(state =>
         selectTradingProviderByNameAndTradeType(
             state,
             tradingDetailContext.trade?.data.exchange,
-            'sell',
+            'exchange',
         ),
     );
 
     return (
         <TradingDetailContext.Provider value={tradingDetailContext}>
             <TradingContainer provider={provider}>
-                <TradingDetailSell />
+                <TradingExchangeDetailContent />
             </TradingContainer>
         </TradingDetailContext.Provider>
     );
