+++ packages/suite/src/views/wallet/trading/exchange/TradingExchangeDetail/TradingExchangeDetailPaymentSending.tsx
@@ -1,4 +1,4 @@
-import { type SellFiatTrade } from 'invity-api';
+import { type ExchangeTrade } from 'invity-api';
 
 import { Translation, type TranslationKey } from '@suite/intl';
 import { formatDurationStrict } from '@suite-common/suite-utils';
@@ -10,34 +10,40 @@ import { Card, Column, InfoItem, type StepListItemState } from '@trezor/componen
 import { useLocales } from 'src/hooks/suite';
 import { useSelector } from 'src/hooks/suite/useSelector';
 import { type Account } from 'src/types/wallet';
+import { TradingDetailStep } from 'src/views/wallet/trading/common/TradingDetail/TradingDetailStep';
+import { TradingDetailTxId } from 'src/views/wallet/trading/common/TradingDetail/TradingDetailTxId';
+import { getTxEstimatedTimeSeconds } from 'src/views/wallet/trading/common/TradingDetail/utils';
 
-import { TradingDetailStep } from '../TradingDetailStep';
-import { TradingDetailTxId } from '../TradingDetailTxId';
-import { getTxEstimatedTimeSeconds } from '../utils';
-
-const getState = (trade: SellFiatTrade): StepListItemState =>
-    trade.status === 'SUCCESS' ? 'done' : 'active';
+const getState = (trade: ExchangeTrade): StepListItemState => {
+    switch (trade.status) {
+        case 'CONVERTING':
+        case 'SUCCESS':
+            return 'done';
+        default:
+            return 'active';
+    }
+};
 
 const getTitleId = (state: StepListItemState): TranslationKey => {
     switch (state) {
         case 'active':
-            return 'TR_SELL_DETAIL_SENDING_TRANSACTION';
+            return 'TR_EXCHANGE_DETAIL_SENDING_TRANSACTION';
         default:
-            return 'TR_SELL_DETAIL_TRANSACTION_SENT';
+            return 'TR_EXCHANGE_DETAIL_TRANSACTION_SENT';
     }
 };
 
-type TradingDetailSellPaymentSendingProps = {
-    trade: SellFiatTrade;
+type TradingExchangeDetailPaymentSendingProps = {
+    trade: ExchangeTrade;
     account?: Account;
     composedTransaction?: TradingComposedTransactionInfo;
 };
 
-export const TradingDetailSellPaymentSending = ({
+export const TradingExchangeDetailPaymentSending = ({
     trade,
     account,
     composedTransaction,
-}: TradingDetailSellPaymentSendingProps) => {
+}: TradingExchangeDetailPaymentSendingProps) => {
     const locale = useLocales();
     const rawFeeInfo = useSelector(state =>
         account ? selectRawNetworkFeeInfo(state, account.symbol) : undefined,
@@ -55,11 +61,11 @@ export const TradingDetailSellPaymentSending = ({
         : undefined;
 
     const txId =
-        trade.txid && account ? (
+        trade.receiveTxHash && account ? (
             <TradingDetailTxId
                 intent="neutral"
                 priority={state === 'done' ? 'secondary' : 'primary'}
-                value={trade.txid}
+                value={trade.receiveTxHash}
                 account={account}
             />
         ) : null;
