+++ packages/suite/src/views/wallet/trading/sell/TradingSellDetail/TradingSellDetail.tsx
@@ -3,25 +3,26 @@ import { selectTradingProviderByNameAndTradeType } from '@suite-common/trading';
 import { useSelector } from 'src/hooks/suite';
 import { TradingDetailContext, useTradingDetail } from 'src/hooks/wallet/trading/useTradingDetail';
 import { TradingContainer } from 'src/views/wallet/trading/common/TradingContainer';
-import { TradingDetailBuy } from 'src/views/wallet/trading/common/TradingDetail/TradingDetailBuy/TradingDetailBuy';
 
-export const TradingBuyDetail = () => {
+import { TradingSellDetailContent } from './TradingSellDetailContent';
+
+export const TradingSellDetail = () => {
     const tradingDetailContext = useTradingDetail({
-        tradeType: 'buy',
+        tradeType: 'sell',
     });
 
     const provider = useSelector(state =>
         selectTradingProviderByNameAndTradeType(
             state,
             tradingDetailContext.trade?.data.exchange,
-            'buy',
+            'sell',
         ),
     );
 
     return (
         <TradingDetailContext.Provider value={tradingDetailContext}>
             <TradingContainer provider={provider}>
-                <TradingDetailBuy />
+                <TradingSellDetailContent />
             </TradingContainer>
         </TradingDetailContext.Provider>
     );
